refactor: move business logic into `enrollment` module
make minimally needed changes to keep unit tests passing. they will
be refactored to separate the view logic from the business logic in a
separate commit.
